backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
praw==7.7.1
nltk==3.8.1
websockets==12.0
python-dotenv==1.0.0
aiosqlite==0.19.0
